rename fuzz_analysis to analysis
